[Refactor] refactoring for PDP
